"Added Houses feature and updated README"
Added new Houses feature to the application. This commit adds new seeder, table, model, controller and API route to handle 'House' data. The README has also been updated to include basic commands for running the application. This was done to expand the real estate offerings of the application and streamline the related instructions.