Update README and Tests for improved clarity
Updated README.md and tests to improve understanding and usage of the project. README now includes detailed installation steps and testing info. Test update includes a new function to test house creation fails with invalid data and improvements in the API request methods to use 'getJson'. This changes improve the readability of the README and the thoroughness of the testing.